Extract options as a separate module
